added storage plus traits
