+++ packages/shade_protocol/Cargo.toml
@@ -25,9 +25,10 @@ errors = []
 flexible_msg = []
 math = []
 storage = ["cosmwasm-storage/iterator"]
+storage_plus = ["dep:secret-storage-plus"]
 
 # Protocol contracts
-airdrop = ["utils", "errors"]
+airdrop = ["utils", "errors", "dep:remain"]
 initializer = ["snip20", "utils"]
 governance = ["utils"]
 mint = ["utils", "snip20"]
@@ -47,12 +48,12 @@ cosmwasm-std = { version = "0.10", package = "secret-cosmwasm-std" }
 cosmwasm-storage = { version = "0.10", package = "secret-cosmwasm-storage" }
 cosmwasm-math-compat = { path = "../cosmwasm_math_compat" }
 cosmwasm-schema = "0.10.1"
-secret-toolkit = { version = "0.2" }
+# secret-toolkit = { version = "0.2" }
 schemars = "0.7"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 snafu = { version = "0.6.3" }
-# Needed for airdrop
-rs_merkle = { git = "https://github.com/FloppyDisck/rs-merkle", branch = "node_export" }
-# Needed for transactions
-query-authentication = { git = "https://github.com/securesecrets/query-authentication", tag = "v1.2.0" }
-remain = "0.2.2"
+# Used by airdrop
+remain = { version = "0.2.2", optional = true }
+# storage plus implementation
+secret-storage-plus = { git = "https://github.com/securesecrets/secret-storage-plus", tag = "v1.0.0", optional = true }
+
