+++ packages/shade_protocol/src/utils/mod.rs
@@ -12,7 +12,6 @@ pub mod flexible_msg;
 #[cfg(feature = "utils")]
 pub mod generic_response;
 
-#[cfg(feature = "storage")]
 pub mod storage;
 
 #[cfg(feature = "math")]
